cmd/govim: track all buffers in govim
Currently we only track .go files in govim. This makes it impossible,
however, to use commands like GOVIMStringFn in non-go files.

Therefore track all buffers within govim, only relaying events to gopls
for files that matter (currently *.go and go.mod files). We also ensure
that we only try and parse *.go files for govim motion support.

As part of this change we more precisely handle events from Vim. The
corresponding methods on *vimstate document this.  Remembering that, per
https://github.com/vim/vim/issues/5336, on Vim startup we do not get
buffer events, we therefore manually trigger govim autocommands to
simulate BufNew, BufCreate and BufRead once govim has finished
initialising.

We also add some context to each fire of an autocommand, such that the
call from Vim -> govim that is made in response to an event includes the
autocommand event definition so that it's easier to spot what event is
firing. e.g:

   recvJSONMsg: ["function","autocommand:3","BufNewFile,BufRead *"...